+++ packages/ui-kit/src/index.ts
@@ -26,6 +26,7 @@ export type {
   OverlayMinimizedPosition,
   OverlayMinimizedVariant,
   PillProps,
+  PillPosition,
   PillVariant,
   TextBlockProps,
 } from './components';